+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
   "editor.formatOnSave": true,
-  "edit.codeActionsOnSave": {
+  "editor.codeActionsOnSave": {
     "source.fixAll": true
   }
 }
