+++ package.json
@@ -17,7 +17,8 @@
   "license": "ISC",
   "dependencies": {
     "@metaplex/js": "^4.11.2",
-    "@project-serum/anchor": "^0.25.0-beta.1",
+    "@solana/spl-token": "^0.3.5",
+    "@solana/web3.js": "^1.66.2",
     "bs58": "^5.0.0",
     "canvas": "^2.9.3",
     "cors": "^2.8.5",
